Add mod mathswe-ts/enum with matching for plain sum type variants
The current matching in the `adt` module is eager which is unsafe since it executes all the map branches automatically with the actual variant type. Only sum types with plain (non fields) matching of variants can be eager since these don't need the variant fields to compute the match result.